+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### next
 - help screen
 - ctrl-c to close a drawer or the help screen
+- 'a' key edits a field, cursor at end, while 'i' puts the cursor at start
 
 <a name="v0.2.0"></a>
 ### v0.2.0 - 2021-09-19